Разбитие css на отдельные модули
