+++ src/Components/Header/Header.css
@@ -1,23 +1,201 @@
 /*--------------------------- Header ------------------------*/
 
 .header {
-    padding: 50px 0 0;
-    font-size: 16px;
-    opacity: .8;
-    letter-spacing: 0px;
+  padding: 50px 0 0;
+  font-size: 16px;
+  opacity: 0.8;
+  letter-spacing: 0px;
+}
+
+.nav_wrapper {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+}
+
+.nav_wrapper a {
+  color: var(--main-text-color);
+  text-decoration: none;
+}
+
+.nav_burger {
+  display: none;
+}
+
+@media (max-width: 1650px) {
+  :root {
+    --section-header-font-size: 70px;
   }
-  
-  .nav_wrapper {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
+
+  .container {
+    max-width: 1300px;
   }
-  
-  .nav_wrapper a {
-    color: var(--main-text-color);
-    text-decoration: none;
+}
+
+@media (max-width: 1440px) {
+  :root {
+    --section-header-font-size: 50px;
   }
-  
-  .nav_burger {
+
+  body {
+    background-image: url(./../../../public/img/bg_1440.svg);
+  }
+
+  .container {
+    max-width: 1000px;
+  }
+
+  .header {
+    padding: 30px 0 0 0;
+    font-size: 14px;
+  }
+}
+
+@media (max-width: 1070px) {
+  :root {
+    --section-header-font-size: 48px;
+  }
+
+  .container {
+    max-width: 961px;
+  }
+}
+
+@media (max-width: 1024px) {
+  body {
+    background-image: url(./../../../public/img/bg_1024.svg);
+  }
+
+  .header {
+    opacity: 1;
+  }
+
+  .nav {
     display: none;
   }
+
+  .nav.open {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 50;
+    width: 100%;
+    height: 100vh;
+    background-color: #120c21;
+  }
+
+  .nav_wrapper.open {
+    display: flex;
+    flex-direction: column;
+    text-align: center;
+    justify-content: space-between;
+    height: 50vh;
+    font-size: 18px;
+  }
+
+  .nav_burger-header {
+    font-size: 18px;
+    font-weight: 700;
+  }
+
+  .nav_burger {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+
+  #burger_menu {
+    z-index: 100;
+    width: 26px;
+    height: 23px;
+    position: relative;
+    transform: rotate(0deg);
+    transition: 0.5s ease-in-out;
+    cursor: pointer;
+  }
+
+  #burger_menu span {
+    display: block;
+    position: absolute;
+    height: 3px;
+    width: 100%;
+    background: #fff;
+    border-radius: 9px;
+    opacity: 1;
+    left: 0;
+    transform: rotate(0deg);
+    transition: 0.25s ease-in-out;
+  }
+
+  #burger_menu span:nth-child(1) {
+    top: 0px;
+  }
+
+  #burger_menu span:nth-child(2),
+  #burger_menu span:nth-child(3) {
+    top: 10px;
+  }
+
+  #burger_menu span:nth-child(4) {
+    top: 20px;
+  }
+
+  #burger_menu.open span:nth-child(1) {
+    top: 18px;
+    width: 0%;
+    left: 50%;
+  }
+
+  #burger_menu.open span:nth-child(2) {
+    transform: rotate(45deg);
+  }
+
+  #burger_menu.open span:nth-child(3) {
+    transform: rotate(-45deg);
+  }
+
+  #burger_menu.open span:nth-child(4) {
+    top: 18px;
+    width: 0%;
+    left: 50%;
+  }
+}
+
+@media (max-width: 1000px) {
+  :root {
+    --section-header-font-size: 30px;
+  }
+
+  .container {
+    max-width: 610px;
+  }
+}
+
+@media (max-width: 680px) {
+  :root {
+    --section-header-font-size: 24px;
+  }
+
+  .container {
+    max-width: 320px;
+  }
+
+  .nav_burger-header {
+    font-size: 16px;
+  }
+}
+
+@media (max-width: 360px) {
+  body {
+    background-image: url(./../../../public/img/bg_360.svg);
+  }
+}
+
+@media (max-width: 340px) {
+  .container {
+    max-width: 293px;
+  }
+}
