+++ src/Components/News/News.css
@@ -1,0 +1,88 @@
+/*--------------------------- News ------------------------*/
+
+.news {
+  margin-top: 282px;
+}
+
+.news_wrapper-header {
+  text-align: center;
+}
+
+.news_wrapper-header h2 {
+  font-size: var(--section-header-font-size);
+}
+
+.news_list {
+  margin: 70px 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 40px;
+}
+
+.news_list-item {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  padding: 23px 32px 23px 38px;
+  box-sizing: border-box;
+  border-radius: 40px;
+  border: 1px rgba(255, 255, 255, 0.7) solid;
+  backdrop-filter: blur(15px);
+}
+
+.title {
+  display: flex;
+  justify-content: space-between;
+  align-items: start;
+}
+
+.title__head {
+  color: white;
+  font-size: 25px;
+  font-weight: 700;
+  word-wrap: break-word;
+}
+
+.title__description {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  display: -moz-box;
+  -moz-box-orient: vertical;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  line-clamp: 2;
+  box-orient: vertical;
+  font-size: 22px;
+  width: 1332px;
+}
+
+.details_adaptive {
+  display: none;
+}
+
+.details {
+  display: flex;
+  flex-direction: column;
+  text-align: right;
+  gap: 2px;
+}
+
+.details__date {
+  color: #fffbfb;
+  font-size: 18px;
+  font-weight: 700;
+  word-wrap: break-word;
+}
+
+.details__theme {
+  font-size: 18px;
+  word-wrap: break-word;
+}
+
+.news_button {
+  text-align: center;
+}
