--- src/Components/School.tsx
@@ -1,22 +1,0 @@
-import React, { FC } from 'react';
-
-import '../App.css';
-
-
-const School: FC = () => {
- 
-
-    return (   
-    <section className="school">
-        <div className="container">
-            <div className="school_wrapper">
-                <div className="school_header"><h2> Студенческая летняя школа по информационной безопасности</h2></div>
-                <p className="school_text">Приглашаем всех желающих студентов кафедры компьютерной инженерии и моделирования принять участие во второй всероссийской студенческой летней школе по информационной безопасности</p>
-                <div className="school_button"><button className="main_description-btn">Подробнее</button></div>
-            </div>
-        </div>
-    </section>
-    );
-  }
-  
-  export default School;
