+++ src/Pages/MainPage.tsx
@@ -4,8 +4,8 @@ import Header from '../Components/Header/Header';
 import Main from '../Components/Main/Main';
 import Milukov from '../Components/Milukov/Milukov';
 import Directions from '../Components/Directions/Directions'
-import News from '../Components/News'
-import School from '../Components/School'
+import News from '../Components/News/News'
+import School from '../Components/School/School'
 import Footer from '../Components/Footer/Footer'
 import '../App.css';
